refactor: вынесена логика DashboardPage в хуки
